moved hashSplitKey function to utils
